Feature Extraction and Similarity Measurement based search engine
All code written in MATLAB - can be implemented using the right libraries in Python.

The idea was to figure out what kind of feature extractor methods and similarity measurement tools work best for different kind of image comparisons.
The project can be expanded and automated using function inputs, and neural networks - not ony to improve retrival accuracy, but to also encompass an automated variety of input image classification.